QA Pass Over `bevy_macro_utils`
Added missing documentation, added `warn(missing_docs)`, and `deny(unsafe)` as standard best-practice.
