+++ crates/bevy_macro_utils/src/attrs.rs
@@ -2,6 +2,7 @@ use syn::{Expr, ExprLit, Lit};
 
 use crate::symbol::Symbol;
 
+/// Get a [literal string](struct@syn::LitStr) from the provided [expression](Expr).
 pub fn get_lit_str(attr_name: Symbol, value: &Expr) -> syn::Result<&syn::LitStr> {
     if let Expr::Lit(ExprLit {
         lit: Lit::Str(lit), ..
@@ -16,6 +17,7 @@ pub fn get_lit_str(attr_name: Symbol, value: &Expr) -> syn::Result<&syn::LitStr>
     }
 }
 
+/// Get a [literal boolean](struct@syn::LitBool) from the provided [expression](Expr) as a [`bool`].
 pub fn get_lit_bool(attr_name: Symbol, value: &Expr) -> syn::Result<bool> {
     if let Expr::Lit(ExprLit {
         lit: Lit::Bool(lit),
