+++ crates/bevy_reflect/bevy_reflect_derive/src/lib.rs
@@ -31,7 +31,6 @@ mod utility;
 
 use crate::derive_data::{ReflectDerive, ReflectMeta, ReflectStruct};
 use crate::type_uuid::gen_impl_type_uuid;
-pub(crate) use bevy_macro_utils::fq_std;
 use container_attributes::ReflectTraits;
 use derive_data::ReflectTypePath;
 use proc_macro::TokenStream;
